Added a lot of admin stuff

+++ src/main/java/com/daymeijroos/iacspring/Admin/AdminProductController.java
@@ -7,6 +7,7 @@ import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.server.ResponseStatusException;
 
@@ -14,41 +15,55 @@ import org.springframework.web.server.ResponseStatusException;
 @RequestMapping(value = "/admin/product")
 public class AdminProductController {
     private final ProductService productService;
+    private final AdminService adminService;
 
     @Autowired
-    public AdminProductController(ProductService productService) {
+    public AdminProductController(ProductService productService, AdminService adminService) {
         this.productService = productService;
+        this.adminService = adminService;
     }
 
-    @RequestMapping(value = "", method = RequestMethod.POST)
-    public ResponseEntity<ProductDTO> postProduct(@RequestBody @Valid ProductDTO productDTO) {
-        return ResponseEntity.ok(productService.post(productDTO));
-    }
-
-    @RequestMapping(value = "", method = RequestMethod.PUT)
-    public ResponseEntity<ProductDTO> updateProduct(@RequestParam String id, @RequestBody ProductDTO productDTO) {
-        try {
-            productDTO.setId(id);
-            return ResponseEntity.ok().body(productService.update(productDTO));
-        } catch (ResourceNotFoundException e) {
-            throw new ResponseStatusException(
-                    HttpStatus.NOT_FOUND, e.getMessage(), e);
+    @PostMapping(value = "")
+    public ResponseEntity<ProductDTO> postProduct(@RequestBody @Valid ProductDTO productDTO, Authentication authentication) {
+        if (!this.adminService.checkForUserId(authentication.getName())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized to add products");
         }
+        return ResponseEntity.ok(this.productService.post(productDTO));
     }
 
-    @RequestMapping(value = "", method = RequestMethod.PATCH, consumes = "application/json-patch+json")
-    public ResponseEntity<ProductDTO> patchProduct(@RequestParam String id, @RequestBody ProductDTO productDTO) {
+    @PutMapping(value = "")
+    public ResponseEntity<ProductDTO> updateProduct(@RequestParam String id, @RequestBody ProductDTO productDTO, Authentication authentication) {
+        if (!this.adminService.checkForUserId(authentication.getName())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized to replace products");
+        }
         productDTO.setId(id);
         try {
-            return ResponseEntity.ok(productService.patch(productDTO));
+            return ResponseEntity.ok().body(this.productService.update(productDTO));
         } catch (ResourceNotFoundException e) {
             throw new ResponseStatusException(
                     HttpStatus.NOT_FOUND, e.getMessage(), e);
         }
     }
 
-    @RequestMapping(value = "", method = RequestMethod.DELETE)
-    public ResponseEntity<String> deleteProduct(@RequestParam String id) {
+    @PatchMapping(value = "", consumes = "application/json-patch+json")
+    public ResponseEntity<ProductDTO> patchProduct(@RequestParam String id, @RequestBody ProductDTO productDTO, Authentication authentication) {
+        if (!this.adminService.checkForUserId(authentication.getName())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized to edit products");
+        }
+        productDTO.setId(id);
+        try {
+            return ResponseEntity.ok(this.productService.patch(productDTO));
+        } catch (ResourceNotFoundException e) {
+            throw new ResponseStatusException(
+                    HttpStatus.NOT_FOUND, e.getMessage(), e);
+        }
+    }
+
+    @DeleteMapping(value = "")
+    public ResponseEntity<String> deleteProduct(@RequestParam String id, Authentication authentication) {
+        if (!this.adminService.checkForUserId(authentication.getName())) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "You are not authorized to remove products");
+        }
         try {
             this.productService.delete(id);
             return ResponseEntity.ok("Product deleted successfully");
